nwjs: Add new loader
ESM scripts in inject_js_start are loaded asynchronously in recent NW.js versions. This causes a race vecause Kawariki modules must be loaded before the application to be able to control the initialization of the latter.
Instead of injecting a userscript, modify the main HTML to load Kawariki modules and only then load all application scripts in their original order (but hoisted to the end of body).
A second window.load event is emulated when all application scripts are done loading.

The new loader is currently used on NW.js versions >= 0.70